new dotfiles

+++ polybar-scripts/aur-updates.sh
@@ -1,12 +1,13 @@
 #!/bin/sh 
 dir="$HOME/polybar-scripts"
 list="$dir/.updatelist"
-#list="/home/joshh/polybar-scripts/.updatelist"
-# script=`realpath $0`     # get absolute path to the script itself
-# exec 6< "$script"        # open bash script using file descriptor 6
-# flock -n 6 || { sleep 5; list="/dev/null"; } # if script is already running, keep going but don't write to file
-#exec 6<>/tmp/aur_lock
-#flock -n 6 || list="/dev/null" # if script is already running, keep going but don't write to file
+polybar_output="$dir/.polybar-aur"
+tmp_lock="/tmp/polybar-aur.lock"
+
+if [ -f $tmp_lock ]; then
+    exit
+fi
+touch $tmp_lock
 
 printf "### Pacman updates ###\n\n" > $list
 
@@ -23,7 +24,9 @@ fi
 updates=$(("$updates_arch" + "$updates_aur"))
 
 if [ "$updates" -gt 0 ]; then
-    echo "$updates_arch  $updates_aur" > "$dir/.polybar-aur"
+    echo "$updates_arch  $updates_aur" > $polybar_output
 else
-    echo "" > "$dir/.polybar-aur"
+    echo "" > $polybar_output
 fi
+
+rm $tmp_lock
